refactor: thin epic-execute-record-wave cli below 500 loc via wave-record-* modules (resolves #1865)
